// src/database/resources/ResourceCatalog.d.ts
export declare const ResourceCatalog: {
    Resource: {
        name: string;
        table_name: string;
    };
    Column: {
        name: string;
        table_name: string;
    };
    Reference: {
        name: string;
        table_name: string;
    };
    Connection: {
        name: string;
        table_name: string;
    };
    ApiAccess: {
        name: string;
        table_name: string;
    };
    Module: {
        name: string;
        table_name: string;
    };
    ModuleMenu: {
        name: string;
        table_name: string;
    };
    ModulePage: {
        name: string;
        table_name: string;
    };
    ResourceAccessPolicy: {
        name: string;
        table_name: string;
    };
    ResourceAccessShare: {
        name: string;
        table_name: string;
    };
    ResourceActivity: {
        name: string;
        table_name: string;
    };
    ResourcePagePermission: {
        name: string;
        table_name: string;
    };
    User: {
        name: string;
        table_name: string;
    };
    UserInfo: {
        name: string;
        table_name: string;
    };
    UserRole: {
        name: string;
        table_name: string;
    };
    Role: {
        name: string;
        table_name: string;
    };
    PersistentLogin: {
        name: string;
        table_name: string;
    };
};